Bug 1498645 - Refactor RichText to its own component
